# rms-api-parent/rms-api-service/target/classes/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Similar Day Service API",
    "version" : "1.0.0",
    "title" : "Similar Day Service API"
  },
  "host" : "localhost",
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json", "application/xml", "application/x-www-form-urlencoded" ],
  "produces" : [ "application/json", "application/xml", "application/x-www-form-urlencoded" ],
  "paths" : {
    "/indexes" : {
      "get" : {
        "tags" : [ "Indexes" ],
        "summary" : "Indexes",
        "description" : "The Indexes endpoint provide all the indexes that were computed in order to automatically select the similar day. The higher an index is, the better it is to use as a similar day\n",
        "parameters" : [ {
          "name" : "date",
          "in" : "query",
          "description" : "Day for which we look at set of 28 days before indexes",
          "required" : true,
          "type" : "string",
          "format" : "date"
        }, {
          "name" : "substation",
          "in" : "query",
          "description" : "Substation for which we look at a indexes",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "An array containing the indexes for the 28 days before the requested date.",
            "schema" : {
              "type" : "object",
              "properties" : { }
            }
          },
          "default" : {
            "description" : "Unexpected error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/residual_load_forecast" : {
      "post" : {
        "tags" : [ "Indexes" ],
        "summary" : "Provides the residual load forecast",
        "description" : "Provides the residual load forecast based on current values of the measurements and current similar day.\n",
        "parameters" : [ {
          "name" : "date",
          "in" : "query",
          "description" : "The date for which we want the residual load forecast.",
          "required" : true,
          "type" : "string",
          "format" : "date"
        }, {
          "name" : "substation",
          "in" : "query",
          "description" : "Substation for which we want the residual laod forecast.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "The residual load forecast",
            "schema" : {
              "$ref" : "#/definitions/Forecast"
            }
          },
          "default" : {
            "description" : "Unexpected error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/run_calculations_request" : {
      "post" : {
        "tags" : [ "Indexes" ],
        "summary" : "Run automatic similar day calculations",
        "description" : "Used to run automatic similar day calculations for the seven next days, including indexes. If similar day have been overrided manually, this new calculations will update the indexes but will NOT update the similar day.\n",
        "parameters" : [ {
          "name" : "date",
          "in" : "query",
          "description" : "We want to compute indexes for the 7 next days after that date.",
          "required" : true,
          "type" : "string",
          "format" : "date"
        } ],
        "responses" : {
          "200" : {
            "description" : "Simple 200 code stating that things went well."
          },
          "default" : {
            "description" : "Unexpected error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/similar_day" : {
      "get" : {
        "tags" : [ "SimilarDay" ],
        "summary" : "Similar Day",
        "description" : "The Similar Day endpoint provide the current similar day for a given substation and a given date\n",
        "parameters" : [ {
          "name" : "date",
          "in" : "query",
          "description" : "Day for which we look at a similar day",
          "required" : true,
          "type" : "string",
          "format" : "date"
        }, {
          "name" : "substation",
          "in" : "query",
          "description" : "Substation for which we look at a similar day",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "The Similar Day",
            "schema" : {
              "$ref" : "#/definitions/SimilarDay"
            }
          },
          "default" : {
            "description" : "Unexpected error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "SimilarDay" ],
        "summary" : "Similar Day Manual Override",
        "description" : "Allows to manually override the similar day used in the system to do the T-forecast\n",
        "parameters" : [ {
          "name" : "date",
          "in" : "query",
          "description" : "Day for which we look at a similar day",
          "required" : true,
          "type" : "string",
          "format" : "date"
        }, {
          "name" : "substation",
          "in" : "query",
          "description" : "Substation for which we look at a similar day",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "newSimilarDay",
          "description" : "New value for the similar day",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SimilarDay"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The Similar Day",
            "schema" : {
              "$ref" : "#/definitions/SimilarDay"
            }
          },
          "default" : {
            "description" : "Unexpected error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "SimilarDay" : {
      "type" : "object",
      "properties" : {
        "similar_day" : {
          "type" : "string",
          "format" : "date",
          "description" : "The similar day (described at midnight UTC)"
        },
        "last_override" : {
          "type" : "string",
          "format" : "date",
          "description" : "The last time a manual override as been made (if not, returns an empty string)"
        }
      }
    },
    "Index" : {
      "type" : "object",
      "properties" : {
        "future" : {
          "type" : "string",
          "format" : "date",
          "description" : "The day for which the indexes are returned (described at midnight UTC)"
        },
        "past" : {
          "type" : "string",
          "format" : "date",
          "description" : "The day which the measurement quality flags were used (described at midnight UTC)"
        },
        "global_index" : {
          "type" : "number",
          "format" : "float",
          "description" : "Global index that summarizes which day is the more adequate to be selected as a similar day by the system"
        },
        "calendar_index" : {
          "type" : "number",
          "format" : "float",
          "description" : "Index used to caracterize whether the day is similar based on calendar data (day of the week, vacations..)"
        },
        "data_quality_index" : {
          "type" : "number",
          "format" : "float",
          "description" : "Index used to caracterize whether the data associated to that day and substation is of good quality."
        }
      }
    },
    "Indexes" : {
      "type" : "object",
      "properties" : {
        "index" : {
          "type" : "array",
          "description" : "Common collection holder for xml binding of Index model",
          "items" : {
            "$ref" : "#/definitions/Index"
          }
        }
      }
    },
    "Forecast" : {
      "type" : "object",
      "properties" : {
        "entity" : {
          "type" : "string",
          "description" : "The name of the substation for the residual load forecast"
        },
        "schedule_type" : {
          "type" : "string",
          "description" : "The nature of the schedule (Residual Load Forecast)"
        },
        "schedule_provider" : {
          "type" : "string",
          "description" : "The value will be \"STEDIN T-Forecast Engine\""
        },
        "values" : {
          "type" : "string",
          "format" : "JSON",
          "description" : "The series based on a specific format [time1,value1], [time2,value2]"
        }
      }
    },
    "Error" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "fields" : {
          "type" : "string"
        }
      }
    }
  }
}